GetTexture expects a uint, and offset over 5500 will overflow a ushort

+++ src/Game/UI/Controls/PaperDollInteractable.cs
@@ -382,13 +382,13 @@ namespace ClassicUO.Game.UI.Controls
                 }
             }
 
-            if (GumpsLoader.Instance.GetTexture((ushort) (animID + offset)) == null)
+            if (GumpsLoader.Instance.GetTexture((uint) (animID + offset)) == null)
             {
                 // inverse
                 offset = isfemale ? Constants.MALE_GUMP_OFFSET : Constants.FEMALE_GUMP_OFFSET;
             }
 
-            if (GumpsLoader.Instance.GetTexture((ushort) (animID + offset)) == null)
+            if (GumpsLoader.Instance.GetTexture((uint) (animID + offset)) == null)
             {
                 Log.Error($"Texture not found in paperdoll: gump_graphic: {(ushort) (animID + offset)}");
             }
